requirements.txt
deepdiff==8.5.0
discord.py==2.5.2
requests==2.32.3
substrate-interface==1.7.11
markdownify==1.1.0
python-dotenv==1.1.0
qrcode==8.2
pillow==11.2.1
aiofiles==24.1.0
